# specs/openapi.yaml
openapi: 3.0.3
info:
  title: music-sync API
  description: Synchronizes playlists between Spotify and YouTube Music
  version: 1.0.0
servers:
  - url: http://localhost/api
    description: Local Docker environment
...
paths:
...
  /auth/spotify/callback:
    get:
      tags: [auth]
      summary: Handle Spotify OAuth2 callback
      operationId: spotifyCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to frontend with ?connected=spotify on success
        "400":
          description: State mismatch or OAuth error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/ytmusic/callback:
    get:
      tags: [auth]
      summary: Handle YouTube Music OAuth2 callback
      operationId: ytmusicCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to frontend with ?connected=ytmusic on success
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/{provider}:
    delete:
      tags: [auth]
      summary: Disconnect a provider
      operationId: disconnectProvider
      parameters:
        - $ref: "#/components/parameters/provider"
      responses:
        "204":
          description: Provider disconnected, tokens deleted
        "404":
          description: Provider was not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─── Playlists ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    provider:
      name: provider
      in: path
      required: true
      schema:
        type: string
        enum: [spotify, ytmusic]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
...